requirements.txt
bcrypt<4.0
boto3
datasets
email-validator
fastapi
ipykernel
jinja2
langchain
langchain_aws
langgraph
matplotlib
notebook
numpy
openai
pandas
passlib
pillow
pydantic[email]
python-dotenv
python-jose
python-multipart
rich
sqlalchemy
tabulate
tqdm
uvicorn
